streamlit/requirements.txt
flask==2.0.1
streamlit==1.22.0
requests==2.28.1
Pillow==9.3.0
numpy==1.23.5
opencv-python==4.7.0.68
librosa==0.9.2
tensorflow==2.11.0
Werkzeug==2.2.2 